+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 10.2.5
 - **FEAT**(callbacks): Add `onStateHistoryChange` callback to `MainEditorCallbacks`.
 - **FEAT**(main-editor): Add `autoCorrectZoomOffset` and `AutoCorrectZoomScale` parameters to the `addLayer` method, allowing layers to be added inside the viewport even when the user is zoomed into a specific area of the editor.
+- **FEAT**(main-editor): Add `closeSubEditor` method to close all subeditors.
 
 ## 10.2.4
 - **FEAT**(layer-interaction): Add `releaseThreshold` to control snapping behavior for helper lines.